rm merge flip_tools into bvec_bval_tools and add robustness to flip_gradients
